+++ hpc/metax/samples/hpcc_helpers.h
@@ -2,4 +2,4 @@
 #pragma once
 
 void gpu_init(int dev_id);
-const char *gpu_name(char *buf, int buf_len);
+const char *gpu_name(int dev_id, char *buf, int buf_len);